Grealy simplify using Maestro
By relaxing how `search` works. For instance, the basis is read
automatically if `:aliases` is missing in the input. The `new` task
function leverages this and is typically the only entry point that users
should use in the vast majority of cases.

Also support a vector of hooks on `:maestro/on-require` in alias data
executed after search.