Introduce Maestro modes
Typically, a repository is organized around well-know way of working.
For instance, testing is a common situation where the same aliases for
tooling and profiles will be invoked no matter what modules are being
tested.

Hence, a mode is typically a keyword pointing to optional aliases and
profiles under `:maestro/mode+` in `deps.edn` and can be activated prior
to search with `:maestro/mode`.