Garantee Maestro selects direct profile for all requested aliases
There was a pathological case where resolving something like `[:a :b]`
where `:b` requires `:a` would result in direct profiles not being
applied in `:a`. While technically correct, this behavior is typically
undesirable.

This patch keeps track of originally requested aliases and activates
direct profile in "indirect" situation if the parent alias was
explicitly requested.