Fix issue with Git libraries having transitive local dependencies
`:local/root` dependencies cannot be compared when consumed externally
via Git dependencies. See: https://clojure.atlassian.net/browse/TDEPS-132

This commit leverages the facto Maestro allows for a monorepo structure
that do not require the use of `:local/root` for development. Hence, it
switches to using Git dependencies all the way. When the repo is stable
and the tree is clean, one can use that commit on the expose task.
Commit the result. That new commit can new be consumed externally and
everything will work out as confirmed by Alex Miller.