Introducing Maestro "uber" modules
The new `$.maestro.uber` namespaces hosts a task for generating
`deps.edn` files that merge all dependencies and paths from a collection
of aliases. All files found in those paths are hard-linked relative to
that generated local `deps.edn` file so that paths from anywhere in the
repository can be consumed without Clojure CLI complaining.

This is actually useful in a very limited set of use cases. In this
repository, it is used for syncing dependencies and paths between the
root `deps.edn` and `bb.edn`.