Greatly simplify requiring Maestro aliases for shell commands
Previously, a `clojure` task was provided that would compute required
aliases and run them with Clojure CLI. However, it was not perfect.

The idea has been generalized in the new `$.maestro.process` namespace
and its `run` task that allows for requiring aliases and templating them
into any arbitrary shell command before running it.